+++ src/main/java/org/mcdealer/mcdealer/ConfigUpdater.java
@@ -4,6 +4,7 @@ import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 
 import java.io.File;
+import java.io.IOException;
 
 public class ConfigUpdater extends MCDealer {
 
@@ -16,10 +17,10 @@ public class ConfigUpdater extends MCDealer {
         FileConfiguration yamlConfig = getConfig();
 
         // Synchronize the values of the configurations
-        syncValues(jsonConfig, yamlConfig);
+        syncValues(yamlConfig, jsonConfig);
 
-        // Save the updated configuration
-        saveConfig();
+        // Save the updated JSON configuration
+        saveJsonConfig(jsonConfig);
     }
 
     private FileConfiguration loadJsonConfig() {
@@ -27,10 +28,19 @@ public class ConfigUpdater extends MCDealer {
         return YamlConfiguration.loadConfiguration(configFile);
     }
 
-    private void syncValues(FileConfiguration jsonConfig, FileConfiguration yamlConfig) {
+    private void syncValues(FileConfiguration yamlConfig, FileConfiguration jsonConfig) {
         // Synchronize the values
-        yamlConfig.set("currencySymbol", jsonConfig.getString("currencySymbol"));
-        yamlConfig.set("currencySymbolPosition", jsonConfig.getString("currencySymbolPosition"));
-        yamlConfig.set("defaultLanguage", jsonConfig.getString("defaultLanguage"));
+        jsonConfig.set("currencySymbol", yamlConfig.getString("currencySymbol"));
+        jsonConfig.set("currencySymbolPosition", yamlConfig.getString("currencySymbolPosition"));
+        jsonConfig.set("defaultLanguage", yamlConfig.getString("defaultLanguage"));
+    }
+
+    private void saveJsonConfig(FileConfiguration jsonConfig) {
+        File configFile = new File(getDataFolder(), "/web/assets/config.json");
+        try {
+            jsonConfig.save(configFile);
+        } catch (IOException e) {
+            this.getLogger().info(" [MCDealer] Json could not be saved ");
+        }
     }
 }